requirements.txt
# Base de dades
psycopg2-binary>=2.9
sqlalchemy>=2.0
pgvector>=0.2
alembic>=1.13

# Ingesta i HTTP
requests>=2.31
python-dotenv>=1.0

# Embeddings locals (sentence-transformers)
# NOTA: torch pot ser >2GB; instal·lar versió CPU-only si no hi ha GPU:
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
sentence-transformers>=2.6
torch>=2.0

# API REST (sessions posteriors)
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.0

# Processament de text
unidecode>=1.3
jellyfish>=1.0

# Scheduler
apscheduler>=3.10

# Tests
pytest>=8.0
pytest-mock>=3.12
